remediation_worker/jobs/azure_network_watcher_network_watcher_enabled/requirements.txt
azure-identity==1.4.1 \
    --hash=sha256:6f95b3505fc134ad16bd16da053456e1933188ac43161704d48ddb4edebf72c9 \
    --hash=sha256:7b071089faf0789059ac24052e311e2b096a002c173d42b96896db09c6e2ba5d
azure-mgmt-resource==16.1.0 \
    --hash=sha256:433cddfb737bbc894698a7c9b07e15b4e0872cf89a54eaf1301d5605b2c4118d \
    --hash=sha256:b814ee27b37f030fe69461ef6f514661340dc8b1f28736362541e1c0d31d90ae
azure-mgmt-network==16.0.0 \
    --hash=sha256:6159a8c44590cc58841690c27c7d4acb0cd9ad0a1e5178c1d35e0f48e3c3c0e9 \
    --hash=sha256:c0e8358e9d530790dbf3efef6b31bce26e664de5096cbd84c62845067da815d1
